fixed for pull request v1.2
